requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
mysql-connector-python==8.0.17
pkg-resources==0.0.0
protobuf==3.9.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
